backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
ollama
huggingface_hub
sqlalchemy
pypdf
python-docx
python-pptx
Pillow
